+++ docs/conf.py
@@ -26,7 +26,7 @@ author = 'Otto Fajardo'
 # The short X.Y version
 version = ''
 # The full version, including alpha/beta/rc tags
-release = '0.3.8'
+release = '0.3.9'
 
 
 # -- General configuration ---------------------------------------------------